+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/helpers/Helper.java
@@ -1,34 +1,23 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.helpers;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.collections4.CollectionUtils;
-import uk.gov.hmcts.ecm.common.helpers.UtilHelper;
 import uk.gov.hmcts.ecm.common.model.bulk.types.DynamicFixedListType;
 import uk.gov.hmcts.ecm.common.model.bulk.types.DynamicValueType;
 import uk.gov.hmcts.ecm.common.model.ccd.CaseData;
 import uk.gov.hmcts.ecm.common.model.ccd.DocumentInfo;
 import uk.gov.hmcts.ecm.common.model.ccd.SignificantItem;
-import uk.gov.hmcts.ecm.common.model.ccd.items.DateListedTypeItem;
-import uk.gov.hmcts.ecm.common.model.ccd.items.HearingTypeItem;
-import uk.gov.hmcts.ecm.common.model.ccd.items.JudgementTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.JurCodesTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.items.RespondentSumTypeItem;
 import uk.gov.hmcts.ecm.common.model.ccd.types.CorrespondenceScotType;
 import uk.gov.hmcts.ecm.common.model.ccd.types.CorrespondenceType;
-import uk.gov.hmcts.ecm.common.model.ccd.types.DateListedType;
 import uk.gov.hmcts.ecm.common.model.labels.LabelPayloadES;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static com.google.common.base.Strings.isNullOrEmpty;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.BF_ACTION_ACAS;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.BF_ACTION_CASE_LISTED;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.BF_ACTION_CASE_PAPERS;
@@ -46,7 +35,6 @@ import static uk.gov.hmcts.ecm.common.model.helper.Constants.BF_ACTION_REPLY_TO_
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.BF_ACTION_STRIKING_OUT_WARNING;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.BRISTOL_CASE_TYPE_ID;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.CASE_CLOSED_POSITION;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.HEARING_STATUS_POSTPONED;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.LEEDS_CASE_TYPE_ID;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.LONDON_CENTRAL_CASE_TYPE_ID;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.LONDON_EAST_CASE_TYPE_ID;
@@ -59,7 +47,6 @@ import static uk.gov.hmcts.ecm.common.model.helper.Constants.NO;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.SCOTLAND_CASE_TYPE_ID;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.WALES_CASE_TYPE_ID;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.WATFORD_CASE_TYPE_ID;
-import static uk.gov.hmcts.ecm.common.model.helper.Constants.YES;
 
 @Slf4j
 public class Helper {
@@ -217,68 +204,11 @@ public class Helper {
                 DynamicListHelper.getDynamicValue(BF_ACTION_STRIKING_OUT_WARNING)));
     }
 
-    public static List<String> hearingMidEventValidation(CaseData caseData) {
-
-        List<String> errors = new ArrayList<>();
-
-        if (caseData.getHearingCollection() != null) {
-            for (HearingTypeItem hearingTypeItem : caseData.getHearingCollection()) {
-                if (hearingTypeItem.getValue().getHearingNumber() == null
-                        || hearingTypeItem.getValue().getHearingNumber().isEmpty()) {
-                    errors.add(HEARING_CREATION_NUMBER_ERROR);
-                    return errors;
-                }
-                if (hearingTypeItem.getValue().getHearingDateCollection() != null) {
-                    for (DateListedTypeItem dateListedTypeItem
-                            : hearingTypeItem.getValue().getHearingDateCollection()) {
-                        if (dateListedTypeItem.getValue().getListedDate() == null
-                                || dateListedTypeItem.getValue().getListedDate().isEmpty()) {
-                            errors.add(HEARING_CREATION_DAY_ERROR);
-                            return  errors;
-                        }
-                    }
-                }
-            }
-        }
-        return errors;
-    }
-
     public static void updatePositionTypeToClosed(CaseData caseData) {
 
         caseData.setPositionType(CASE_CLOSED_POSITION);
         caseData.setCurrentPosition(CASE_CLOSED_POSITION);
 
-    }
-
-    public static void updatePostponedDate(CaseData caseData) {
-
-        if (caseData.getHearingCollection() != null) {
-            for (HearingTypeItem hearingTypeItem : caseData.getHearingCollection()) {
-
-                if (hearingTypeItem.getValue().getHearingDateCollection() != null) {
-                    for (DateListedTypeItem dateListedTypeItem
-                            : hearingTypeItem.getValue().getHearingDateCollection()) {
-
-                        var dateListedType = dateListedTypeItem.getValue();
-                        if (isHearingStatusPostponed(dateListedType) && dateListedType.getPostponedDate() == null) {
-                            dateListedType.setPostponedDate(UtilHelper.formatCurrentDate2(LocalDate.now()));
-                        }
-                        if (dateListedType.getPostponedDate() != null
-                                &&
-                                (!isHearingStatusPostponed(dateListedType)
-                                        || dateListedType.getHearingStatus() == null)) {
-                            dateListedType.setPostponedDate(null);
-                        }
-                    }
-                }
-            }
-        }
-
-    }
-
-    private static boolean isHearingStatusPostponed(DateListedType dateListedType) {
-        return dateListedType.getHearingStatus() != null
-                && dateListedType.getHearingStatus().equals(HEARING_STATUS_POSTPONED);
     }
 
     public static List<String> getAllOffices() {
@@ -332,38 +262,6 @@ public class Helper {
                 .map(jurCodesTypeItem -> jurCodesTypeItem.getValue().getJuridictionCodesList())
                 .collect(Collectors.toList())
                 : new ArrayList<>();
-    }
-
-    public static String findHearingNumber(CaseData caseData, String hearingDate) {
-        if (CollectionUtils.isNotEmpty(caseData.getHearingCollection())) {
-            for (HearingTypeItem hearingTypeItem : caseData.getHearingCollection()) {
-                for (DateListedTypeItem dateListedTypeItem : hearingTypeItem.getValue().getHearingDateCollection()) {
-                    var listedDate = dateListedTypeItem.getValue().getListedDate().substring(0, 10);
-                    if (listedDate.equals(hearingDate)) {
-                        return hearingTypeItem.getValue().getHearingNumber();
-                    }
-                }
-            }
-        }
-        return null;
-    }
-
-    public static void populateJudgmentDateOfHearing(CaseData caseData) throws ParseException {
-        if (CollectionUtils.isNotEmpty(caseData.getJudgementCollection())) {
-            for (JudgementTypeItem judgementTypeItem : caseData.getJudgementCollection()) {
-                if (NO.equals(judgementTypeItem.getValue().getNonHearingJudgment())) {
-                    var hearingDate = judgementTypeItem.getValue().getDynamicJudgementHearing().getValue().getLabel();
-                    hearingDate = hearingDate.substring(hearingDate.length() - 11);
-                    var simpleDateFormat = new SimpleDateFormat("dd MMM yyyy");
-                    var date = simpleDateFormat.parse(hearingDate);
-                    simpleDateFormat.applyPattern("yyyy-MM-dd");
-                    judgementTypeItem.getValue().setJudgmentHearingDate(simpleDateFormat.format(date));
-                } else {
-                    judgementTypeItem.getValue().setDynamicJudgementHearing(null);
-                    judgementTypeItem.getValue().setJudgmentHearingDate(null);
-                }
-            }
-        }
     }
 
 }
